requirements.txt
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
keyboard==0.13.5
ollama==0.6.1
pillow==12.2.0
pydantic==2.13.1
pydantic_core==2.46.1
pystray==0.19.5
six==1.17.0
typing-inspection==0.4.2
typing_extensions==4.15.0
